perf: 依据Ob官方整改意见调整代码
Copyright (c) 2023 ZhangGithub
Wrong year.

"description": "同步ob笔记到博客园",
Rewrite your description so that it does not include "OB", which I interpret to be an abreviation of "Obsidian"
Avoid including sentences like "This is a plugin for Obsidian that does ..." in your description, it should be a given since this is only shown when browsing plugins through the app or on our website.
We have published a guide for plugin descriptions to follow here: https://docs.obsidian.md/Plugins/Releasing/Submission+requirements+for+plugins#Keep+plugin+descriptions+short+and+simple.

new Notice("Goodbye !")
Please remove this.

blog_url, location_posts, throttling_mode
Please use proper names for the settings.

const {vault} = this.app;, const {metadataCache} = this.app;, const {metadataCache} = this.app;
It's highly discouraged that you use the global app instance in your plugin. It's currently accessible from the global context for debugging purposes. Instead, you should use the reference to App that you have from your plugin instance.

while (imgLink.startsWith("../")) {
All this code should be replaced with a call to the getFirstLinkpathDest function,